requirements.txt
Django==3.2.6
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.29
psycopg2==2.9.1
pytest==6.2.5
pytest-django==4.4.0
pytest-sugar==0.9.4
